+++ phpunit.xml
@@ -3,8 +3,8 @@
          colors="true">
     <testsuites>
         <testsuite name="unit">
-            <directory suffix="Test.php">./tests/unit/</directory>
-            <exclude>tests/unit/AbstractLoaderTest.php</exclude>
+            <directory suffix="Test.php">./tests/</directory>
+            <exclude>tests/AbstractLocatorTest.php</exclude>
         </testsuite>
     </testsuites>
