+++ tests/container/ServiceTest.php
@@ -1,7 +1,6 @@
 <?php
-namespace smpl\mydi\tests\unit\container;
+namespace smpl\mydi\container;
 
-use smpl\mydi\container\Service;
 use smpl\mydi\LocatorInterface;
 
 class ServiceTest extends \PHPUnit_Framework_TestCase
