+++ tests/loader/executor/LazyFactoryTest.php
@@ -1,24 +1,26 @@
 <?php
-namespace smpl\mydi\tests\unit\loader\executor;
+namespace smpl\mydi\loader\executor;
 
-use smpl\mydi\loader\executor\Service;
+use smpl\mydi\container\Lazy;
 use smpl\mydi\LocatorInterface;
 
-class ServiceTest extends \PHPUnit_Framework_TestCase
+class LazyFactoryTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @var Service
+     * @var LazyFactory
      */
     protected $executor;
-    protected $wrapperResult = \stdClass::class;
 
     public function testExecute()
     {
         $locator = $this->getMock(LocatorInterface::class);
         /** @var LocatorInterface $locator */
         $result = $this->executor->execute('stdClass', []);
-        $this->assertInstanceOf(\smpl\mydi\container\Service::class, $result);
-        $this->assertInstanceOf(\stdClass::class, $result->resolve($locator));
+        $this->assertInstanceOf(Lazy::class, $result);
+        $this->assertInstanceOf(\Closure::class, $result->resolve($locator));
+        $value1 = $result->resolve($locator);
+        $value2 = $result->resolve($locator);
+        $this->assertNotSame($value1(), $value2());
     }
 
     public function testExecuteWithString()
@@ -31,8 +33,9 @@ class ServiceTest extends \PHPUnit_Framework_TestCase
         /** @var LocatorInterface $locator */
 
         $result = $this->executor->execute('ololo', \stdClass::class);
-        $this->assertInstanceOf(\smpl\mydi\container\Service::class, $result);
-        $this->assertSame(1234, $result->resolve($locator));
+        $this->assertInstanceOf(Lazy::class, $result);
+        $result = $result->resolve($locator);
+        $this->assertSame(1234, $result());
     }
 
     /**
@@ -67,6 +70,6 @@ class ServiceTest extends \PHPUnit_Framework_TestCase
     protected function setUp()
     {
         parent::setUp();
-        $this->executor = new Service();
+        $this->executor = new LazyFactory();
     }
 }
